Revert "fix #59: use Integer values for variable indices"
This reverts commit 923ead0adca073344cdda99cd3b761b0b3129c32.

Someting's wrong and I mistakenly tested the changes with an outdated version.

+++ src/Theory/Proof/CaseDistinctions.hs
@@ -296,7 +296,7 @@ precomputeCaseDistinctions ctxt typAsms =
 
     absFact (Fact tag ts) = (tag, length ts)
 
-    nMsgVars n = [ varTerm (LVar "t" LSortMsg i) | i <- [1..fromIntegral n] ]
+    nMsgVars n = [ varTerm (LVar "t" LSortMsg i) | i <- [1..n] ]
 
     someProtoGoal :: (FactTag, Int) -> LNFact
     someProtoGoal (tag, arity) = Fact tag (nMsgVars arity)
